Allow multiple link targets
link_program() produces a single output according to the extension
defined in OUPUT_EXT ("bin", "elf", "hex", etc.).  This commit
allows that configured entensions to be a list.  In the case of a
list, an output will be produced for each of the extensions in the
list, e.g. setting OUTPUT_EXT to ["bin", "hex] will result in the
output of both .bin and .hex (Intel hex) formats.

Signed-off-by: Kyle Stein <kyle.stein@arm.com>